⚡Discrete Scale uses built-in plotly vis
